Moved cosmosdb specific logic out of report manager and into cosmos collection
